Finished tutorial
Problems with docker came up that, after 5 hours, seem unfixable for seemingly no reason. So last 10 minutes of tutorial had to be skipped where we run docker commands and have run the sight through docker with a database as a separate entity. Hopefully, I can resolve this issue soon, but for the meantime, this project is complete.